feat: 에러 처리 로직
