+++ frontend/src/custom-error/index.ts
@@ -1,25 +1,2 @@
-/* eslint-disable max-classes-per-file */
-
-export class CustomError extends Error {
-  constructor(message: string) {
-    super(message);
-    this.name = this.constructor.name;
-  }
-}
-
-export class NetworkError extends CustomError {
-  constructor(message: string = "아무런 응답을 받지 못했습니다") {
-    super(message);
-  }
-}
-
-export class UnknownError extends CustomError {
-  constructor(
-    message: string = "알 수 없는 에러가 발생했습니다. 다시 시도해 주세요",
-  ) {
-    super(message);
-  }
-  static getMessage(): string {
-    return "알 수 없는 에러가 발생했습니다. 다시 시도해 주세요";
-  }
-}
+export * from "./api-error";
+export * from "./custom-error";
